Add stacked bar chart

+++ src/Chart.php
@@ -21,7 +21,7 @@ class Chart
     public array $yAxis = [];
 
     /**
-     * @param  Element[]  $series
+     * @param  Serie[]  $series
      * @param  Renderable[]  $annotations
      * @param  YAxis|YAxis[]  $yAxis
      */
@@ -38,10 +38,10 @@ class Chart
         public XAxis $xAxis = new XAxis,
         public array $annotations = [],
         public array $series = [],
-        protected int $leftMargin = 10,
-        protected int $rightMargin = 30,
-        protected int $bottomMargin = 50,
-        protected int $topMargin = 25,
+        protected float $leftMargin = 10,
+        protected float $rightMargin = 30,
+        protected float $bottomMargin = 50,
+        protected float $topMargin = 25,
     ) {
         $this->yAxis = is_array($yAxis) ? $yAxis : [$yAxis];
         $this->yAxis = array_reduce($this->yAxis, function (array $carry, YAxis $yAxis) {
